+++ doc/CHANGELOG.md
@@ -182,6 +182,7 @@
   - Markup has changed to a custom element `<ids-tree><ids-tree-node></ids-tree-node></ids-tree>`
   - If using events, events are now plain JS events.
   - Can now be imported as a single JS file and used with encapsulated styles
+- `[Treemap]` The Treemap component has been changed to a web component and used with ids-treemap
 - `[Wizard]` The Wizard component has been changed to a web component. ([#5026](https://github.com/infor-design/enterprise/issues/5026))
   - Wizard is now a custom element `<ids-wizard></ids-wizard>`
   - Steps are now codified as custom elements `<ids-wizard-step>MY-LABEL</ids-wizard-step>`